Pawn promotion and initialized multiplayer scene
Fixed the code for pawn promotion so the pawn is promoted to a queen once it reaches the other side of the board. Can add a prompt later on so the player can choose what piece they want to promote the pawn to. Also initialized the scene for offline multiplayer and added navigation